Trait fedimint_core_client::ModuleCommon

pub trait ModuleCommon {
    fn module_key() -> u16;
    fn decode_spendable_output(
        r: &mut dyn Read
    ) -> Result<SpendableOutput, DecodeError>;
    fn decode_input(r: &mut dyn Read) -> Result<Input, DecodeError>;
    fn decode_output(r: &mut dyn Read) -> Result<Output, DecodeError>;
    fn decode_pending_output(
        r: &mut dyn Read
    ) -> Result<PendingOutput, DecodeError>;
    fn decode_output_outcome(
        r: &mut dyn Read
    ) -> Result<OutputOutcome, DecodeError>;
}

Common functionality of a Fedimint module

Both backend and server part of the module will need things like decoding module-specific data.

Required Methods

fn module_key() -> u16

fn decode_spendable_output(
r: &mut dyn Read
) -> Result<SpendableOutput, DecodeError>

Decode SpendableOutput compatible with this module, after the module key prefix was already decoded

fn decode_input(r: &mut dyn Read) -> Result<Input, DecodeError>

Decode Input compatible with this module, after the module key prefix was already decoded

fn decode_output(r: &mut dyn Read) -> Result<Output, DecodeError>

Decode Output compatible with this module, after the module key prefix was already decoded

fn decode_pending_output(r: &mut dyn Read) -> Result<PendingOutput, DecodeError>

Decode PendingOutput compatible with this module, after the module key prefix was already decoded

fn decode_output_outcome(r: &mut dyn Read) -> Result<OutputOutcome, DecodeError>

Decode OutputOutcome compatible with this module, after the module key prefix was already decoded